+++ vpin-studio-rest-client/src/main/java/de/mephisto/vpin/restclient/system/SystemId.java
@@ -1,0 +1,35 @@
+package de.mephisto.vpin.restclient.system;
+
+public class SystemId {
+  private String systemId;
+
+  private String systemName;
+
+  private String version;
+
+
+  public String getSystemId() {
+    return systemId;
+  }
+
+  public void setSystemId(String systemId) {
+    this.systemId = systemId;
+  }
+
+  public String getSystemName() {
+    return systemName;
+  }
+
+  public void setSystemName(String systemName) {
+    this.systemName = systemName;
+  }
+
+  public String getVersion() {
+    return version;
+  }
+
+  public void setVersion(String version) {
+    this.version = version;
+  }
+
+}